restructured file structure to better fit MVVM and more clearly describe purpose
