--- app/src/main/java/com/example/mediasoundfilter/MainActivity.kt
@@ -1,28 +1,0 @@
-package com.example.mediasoundfilter
-
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
-import androidx.compose.ui.Modifier
-import com.example.mediasoundfilter.screens.login.LoginScreen
-import com.example.mediasoundfilter.ui.theme.MediaSoundFilterTheme
-
-class MainActivity : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContent {
-            MediaSoundFilterTheme {
-                // A surface container using the 'background' color from the theme
-                Surface(
-                    modifier = Modifier.fillMaxSize(),
-                    color = MaterialTheme.colorScheme.background
-                ) {
-                    MediaSoundFilterApp()
-                }
-            }
-        }
-    }
-}
