+++ app/src/main/java/com/example/mediasoundfilter/data/responsemodel/VideoList.kt
@@ -1,0 +1,9 @@
+package com.example.mediasoundfilter.data.responsemodel
+
+import com.squareup.moshi.Json
+import com.squareup.moshi.JsonClass
+
+@JsonClass(generateAdapter = true)
+data class VideoList (
+    @Json(name = "items") val items: List<VideoDTO> = emptyList()
+)
