+++ app/src/main/java/com/example/mediasoundfilter/domain/model/Video.kt
@@ -1,0 +1,6 @@
+package com.example.mediasoundfilter.domain.model
+
+data class Video (
+    val title: String = "",
+    val channelTitle: String = ""
+)
