--- app/src/main/java/com/example/mediasoundfilter/model/Video.kt
@@ -1,6 +1,0 @@
-package com.example.mediasoundfilter.model
-
-data class Video (
-    val title: String = "",
-    val channelTitle: String = ""
-)
