--- app/src/main/java/com/example/mediasoundfilter/responsemodel/VideoDTO.kt
@@ -1,10 +1,0 @@
-package com.example.mediasoundfilter.responsemodel
-
-import com.squareup.moshi.Json
-import com.squareup.moshi.JsonClass
-
-@JsonClass(generateAdapter = true)
-class VideoDTO {
-    @Json(name = "snippet.title") val title: String = ""
-    @Json(name = "snippet.channelTitle") val channelTitle: String = ""
-}
