+++ app/src/main/java/com/example/mediasoundfilter/ui/Constants.kt
@@ -1,0 +1,3 @@
+package com.example.mediasoundfilter.ui
+
+const val MIN_PASSWORD_LENGTH = 6
