+++ app/src/main/java/com/example/mediasoundfilter/ui/screens/upload/UploadUiState.kt
@@ -1,0 +1,6 @@
+package com.example.mediasoundfilter.ui.screens.upload
+
+data class UploadUiState (
+    val videoId: String? = null,
+    val linkError: String? = null
+)
