--- app/src/main/java/com/example/mediasoundfilter/uistate/UploadUiState.kt
@@ -1,6 +1,0 @@
-package com.example.mediasoundfilter.uistate
-
-data class UploadUiState (
-    val videoId: String? = null,
-    val linkError: String? = null
-)
