--- app/src/main/java/com/example/mediasoundfilter/viewmodel/Constants.kt
@@ -1,3 +1,0 @@
-package com.example.mediasoundfilter.viewmodel
-
-const val MIN_PASSWORD_LENGTH = 6
